TIP-613: Extract ProductUniqueData content in abstract class
